intuitiveobjects-rag-server/requirements.txt
acres==0.5.0
aio-pika==9.5.5
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiormq==6.8.1
aiosignal==1.3.2
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chromadb==1.0.13
ci-info==0.3.0
click==8.1.8
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
dataclasses-json==0.6.7
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
docling==2.39.0
docling-core==2.39.0
docling-ibm-models==3.6.0
docling-parse==4.1.0
durationpy==0.10
easyocr==1.7.2
email_validator==2.2.0
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.3.0
fastapi==0.115.8
fastapi-mail==1.5.0
filelock==3.18.0
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.26
langchain-core==0.3.66
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langsmith==0.4.4
latex2mathml==3.78.0
lazy_loader==0.4
looseversion==1.3.0
lxml==5.4.0
Markdown==3.8.2
markdown-it-py==3.0.0
markdownify==1.1.0
marko==2.1.4
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
motor==3.7.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.5.1
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.5
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
numpy==2.3.1
oauthlib==3.2.2
ollama==0.5.1
onnxruntime==1.22.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pamqp==3.3.0
pandas==2.3.0
pathlib==1.0.1
pillow==11.2.1
pluggy==1.6.0
posthog==2.4.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
pydot==4.0.1
Pygments==2.19.2
PyJWT==2.10.1
pylatexenc==2.10
pymongo==4.11.1
PyMuPDF==1.26.1
pyparsing==3.2.3
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-engineio==4.12.2
python-multipart==0.0.20
python-pptx==1.0.2
python-socketio==5.13.0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
semchunk==2.2.2
sentence-transformers==4.1.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.45.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.6.11
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xlsxwriter==3.2.5
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
